requirements.txt
asgiref>=3.8,<3.9
Django==5.2.7
sqlparse==0.5.3
tzdata==2025.2
python-dotenv==1.0.0
# requirements.txt
django-ratelimit==4.1.0

# Document Verification (OCR)
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3